chore(refactor): ZubZet class to use all functionality as traits
