+++ src/Message/RequestResponseHandler.php
@@ -2,13 +2,15 @@
 
     namespace ZubZet\Framework\Message;
 
-    use ZubZet\Framework\Core\Model;
     use ZubZet\Framework\ZubZet;
+    use ZubZet\Framework\Core\CanRetrieveModel;
 
     /**
      * Base class for the response and request objects
      */
     class RequestResponseHandler {
+
+        use CanRetrieveModel;
 
         /**
          * @var ZubZet $booter The framework object
@@ -48,35 +50,21 @@
         }
 
         /**
-         * Gets the database communication interface
-         * @return Model
-         */
-        public function getModel() {
-            return model(...func_get_args());
-        }
-
-        /**
          * Gets a booter setting
          * @param string $key Key of the setting
          * @return mixed Value of the key
          */
         public function getBooterSettings($key = null, $useDefault = true, $default = null) {
             // Return all settings if no key is provided
-            if(empty($key)) {
-                return $this->booter->configuration->settings;
-            }
+            if(empty($key)) return zubzet()->getAllAttributes();
 
             // Return specific setting if it exists
-            if(isset($this->booter->configuration->{$key})) {
-                return $this->booter->configuration->{$key};
-            }
+            if(isset(zubzet()->{$key})) return zubzet()->{$key};
 
             // Return default if enabled
-            if($useDefault) {
-                return $default;
-            }
+            if($useDefault) return $default;
 
-            throw new \Exception("The setting '$key' does not exist!");
+            throw new \InvalidArgumentException("The setting '$key' does not exist!");
         }
     }
 ?>
