+++ src/Message/StateManagement/Input.php
@@ -1,0 +1,10 @@
+<?php
+
+    namespace ZubZet\Framework\Message\StateManagement;
+
+
+    class Input {
+
+    }
+
+?>
