+++ src/Message/StateManagement/Output.php
@@ -1,0 +1,9 @@
+<?php
+
+    namespace ZubZet\Framework\Message\StateManagement;
+
+    class Output {
+
+    }
+
+?>
