+++ tests/e2e/tests/cypress.config.js
@@ -4,7 +4,7 @@ module.exports = defineConfig({
   e2e: {
     baseUrl: "http://localhost:8080",
     experimentalRunAllSpecs: true,
-    video: true,
+    video: false,
     viewportWidth: 1280,
     viewportHeight: 720,
   },
